rename cawp_time to cawp
