--- config/data_sources/cawp_time_data.tf
@@ -1,7 +1,0 @@
-# Resources and routines for CAWP TIME Data ingestion.
-
-# Create a BigQuery dataset for CAWP TIME data.
-resource "google_bigquery_dataset" "cawp_time" {
-  dataset_id = "cawp_time_data"
-  location   = "US"
-}
