+++ data/cawp/cawp_state_leg_69.csv
@@ -1,0 +1,49 @@
+time_period,state_fips,total_state_leg_count
+"1978","69",23
+"1979","69",23
+"1980","69",23
+"1981","69",23
+"1982","69",23
+"1983","69",23
+"1984","69",24
+"1985","69",24
+"1986","69",25
+"1987","69",25
+"1988","69",23
+"1989","69",23
+"1990","69",23
+"1991","69",23
+"1992","69",27
+"1993","69",27
+"1994","69",29
+"1995","69",29
+"1996","69",26
+"1997","69",26
+"1998","69",27
+"1999","69",27
+"2000","69",27
+"2001","69",27
+"2002","69",27
+"2003","69",27
+"2004","69",27
+"2005","69",27
+"2006","69",27
+"2007","69",27
+"2008","69",29
+"2009","69",29
+"2010","69",29
+"2011","69",29
+"2012","69",29
+"2013","69",29
+"2014","69",29
+"2015","69",29
+"2016","69",29
+"2017","69",29
+"2018","69",29
+"2019","69",29
+"2020","69",29
+"2021","69",29
+"2022","69",29
+"2023","69",29
+"2024","69",29
+"2025","69",29
