--- data/cawp_time/cawp_state_leg_11.csv
@@ -1,54 +1,0 @@
-time_period,state_fips,total_state_leg_count
-"1973","11",13
-"1974","11",13
-"1975","11",13
-"1976","11",13
-"1977","11",13
-"1978","11",13
-"1979","11",13
-"1980","11",13
-"1981","11",13
-"1982","11",13
-"1983","11",13
-"1984","11",13
-"1985","11",13
-"1986","11",13
-"1987","11",13
-"1988","11",13
-"1989","11",13
-"1990","11",13
-"1991","11",13
-"1992","11",13
-"1993","11",13
-"1994","11",13
-"1995","11",13
-"1996","11",13
-"1997","11",13
-"1998","11",13
-"1999","11",13
-"2000","11",13
-"2001","11",13
-"2002","11",13
-"2003","11",13
-"2004","11",13
-"2005","11",13
-"2006","11",13
-"2007","11",13
-"2008","11",13
-"2009","11",13
-"2010","11",13
-"2011","11",13
-"2012","11",13
-"2013","11",13
-"2014","11",13
-"2015","11",13
-"2016","11",13
-"2017","11",13
-"2018","11",13
-"2019","11",13
-"2020","11",13
-"2021","11",13
-"2022","11",13
-"2023","11",13
-"2024","11",13
-"2025","11",13
