--- data/cawp_time/cawp_state_leg_66.csv
@@ -1,77 +1,0 @@
-time_period,state_fips,total_state_leg_count
-"1950","66",21
-"1951","66",21
-"1952","66",21
-"1953","66",21
-"1954","66",21
-"1955","66",21
-"1956","66",21
-"1957","66",21
-"1958","66",21
-"1959","66",21
-"1960","66",21
-"1961","66",21
-"1962","66",21
-"1963","66",21
-"1964","66",21
-"1965","66",21
-"1966","66",21
-"1967","66",21
-"1968","66",21
-"1969","66",21
-"1970","66",21
-"1971","66",21
-"1972","66",21
-"1973","66",21
-"1974","66",21
-"1975","66",21
-"1976","66",21
-"1977","66",21
-"1978","66",21
-"1979","66",21
-"1980","66",21
-"1981","66",21
-"1982","66",21
-"1983","66",21
-"1984","66",21
-"1985","66",21
-"1986","66",21
-"1987","66",21
-"1988","66",21
-"1989","66",21
-"1990","66",21
-"1991","66",21
-"1992","66",21
-"1993","66",21
-"1994","66",21
-"1995","66",21
-"1996","66",15
-"1997","66",15
-"1998","66",15
-"1999","66",15
-"2000","66",15
-"2001","66",15
-"2002","66",15
-"2003","66",15
-"2004","66",15
-"2005","66",15
-"2006","66",15
-"2007","66",15
-"2008","66",15
-"2009","66",15
-"2010","66",15
-"2011","66",15
-"2012","66",15
-"2013","66",15
-"2014","66",15
-"2015","66",15
-"2016","66",15
-"2017","66",15
-"2018","66",15
-"2019","66",15
-"2020","66",15
-"2021","66",15
-"2022","66",15
-"2023","66",15
-"2024","66",15
-"2025","66",15
