--- data/cawp_time/cawp_state_leg_72.csv
@@ -1,75 +1,0 @@
-time_period,state_fips,total_state_leg_count
-"1952","72",78
-"1953","72",78
-"1954","72",78
-"1955","72",78
-"1956","72",78
-"1957","72",78
-"1958","72",78
-"1959","72",78
-"1960","72",78
-"1961","72",78
-"1962","72",78
-"1963","72",78
-"1964","72",78
-"1965","72",78
-"1966","72",78
-"1967","72",78
-"1968","72",78
-"1969","72",78
-"1970","72",78
-"1971","72",78
-"1972","72",78
-"1973","72",78
-"1974","72",78
-"1975","72",78
-"1976","72",78
-"1977","72",78
-"1978","72",78
-"1979","72",78
-"1980","72",78
-"1981","72",78
-"1982","72",78
-"1983","72",78
-"1984","72",78
-"1985","72",78
-"1986","72",78
-"1987","72",78
-"1988","72",78
-"1989","72",78
-"1990","72",78
-"1991","72",78
-"1992","72",78
-"1993","72",78
-"1994","72",78
-"1995","72",78
-"1996","72",78
-"1997","72",78
-"1998","72",78
-"1999","72",78
-"2000","72",78
-"2001","72",78
-"2002","72",78
-"2003","72",78
-"2004","72",78
-"2005","72",78
-"2006","72",78
-"2007","72",78
-"2008","72",78
-"2009","72",78
-"2010","72",78
-"2011","72",78
-"2012","72",78
-"2013","72",78
-"2014","72",78
-"2015","72",78
-"2016","72",78
-"2017","72",78
-"2018","72",78
-"2019","72",78
-"2020","72",78
-"2021","72",78
-"2022","72",78
-"2023","72",78
-"2024","72",78
-"2025","72",78
