--- data/cawp_time/cawp_state_leg_78.csv
@@ -1,73 +1,0 @@
-time_period,state_fips,total_state_leg_count
-"1954","78",11
-"1955","78",11
-"1956","78",11
-"1957","78",11
-"1958","78",11
-"1959","78",11
-"1960","78",11
-"1961","78",11
-"1962","78",11
-"1963","78",11
-"1964","78",11
-"1965","78",11
-"1966","78",15
-"1967","78",15
-"1968","78",15
-"1969","78",15
-"1970","78",15
-"1971","78",15
-"1972","78",15
-"1973","78",15
-"1974","78",15
-"1975","78",15
-"1976","78",15
-"1977","78",15
-"1978","78",15
-"1979","78",15
-"1980","78",15
-"1981","78",15
-"1982","78",15
-"1983","78",15
-"1984","78",15
-"1985","78",15
-"1986","78",15
-"1987","78",15
-"1988","78",15
-"1989","78",15
-"1990","78",15
-"1991","78",15
-"1992","78",15
-"1993","78",15
-"1994","78",15
-"1995","78",15
-"1996","78",15
-"1997","78",15
-"1998","78",15
-"1999","78",15
-"2000","78",15
-"2001","78",15
-"2002","78",15
-"2003","78",15
-"2004","78",15
-"2005","78",15
-"2006","78",15
-"2007","78",15
-"2008","78",15
-"2009","78",15
-"2010","78",15
-"2011","78",15
-"2012","78",15
-"2013","78",15
-"2014","78",15
-"2015","78",15
-"2016","78",15
-"2017","78",15
-"2018","78",15
-"2019","78",15
-"2020","78",15
-"2021","78",15
-"2022","78",15
-"2023","78",15
-"2024","78",15
-"2025","78",15
