--- python/datasources/cawp_time.py
@@ -1,1055 +1,0 @@
-"""
-Data Source: CAWP (data on women in elected office)
-
-Description:
-- The data on women by race by year by office is downloaded from the CAWP website.
-- The downloaded data is stored locally in /data/cawp_time
-
-Instructions for Downloading NUMERATOR Data:
-1. Visit the CAWP Database by Race website:
-    https://cawpdata.rutgers.edu/women-elected-officials/race-ethnicity
-2. Under "Filter By Date", select "Show All Years"
-3. Under "Level of Office", select "Congress", "State Legislative", and "Territorial/D.C."
-4. Click the "Search" button at the bottom to refresh the page with the new selections
-5. Once the page loads, click "Download" button
-6. In the modal that pops up, log in with valid account info (it's free)
-7. Once logged in, click the "Download" button again
-8. In the next modal, download, slowly, as a .csv file
-9. Once exported, save (renaming) as /data/cawp_time/cawp-by_race_and_ethnicity_time_series.csv
-10. Commit to the repo, and if needed rerun the STAGING (infra TEST) and PROD DAG pipelines
-
-Notes:
-- This is simply the "numerator" data; we rely on directly downloaded table data from CAWP
- for state legislature demoninators, and UnitedStates.io for Congress denominators
-
-Instructions for Updating TERRITORIAL LEGISLATURE DENOMINATOR Data:
-- Once per year, we should also update our custom-made denominator files for territorial leg
-- These are the other .csv files in that same /data/cawp_time directory, named by their FIPS code.
-- Data source: ncsl.org/resources/details/number-of-legislators-and-length-of-terms-in-years
-
-Last Updated: Feb. 2024
-"""
-
-import pandas as pd
-from ingestion.merge_utils import ACS_EARLIEST_YEAR, ACS_CURRENT_YEAR
-from ingestion.standardized_columns import Race
-from ingestion.dataset_utils import (
-    generate_pct_rel_inequity_col,
-    zero_out_pct_rel_inequity,
-    preserve_only_current_time_period_rows,
-)
-from ingestion import gcs_to_bq_util, merge_utils
-import ingestion.standardized_columns as std_col
-from datasources.data_source import DataSource
-from ingestion.constants import (
-    NATIONAL_LEVEL,
-    STATE_LEVEL,
-    STATE_LEVEL_FIPS_LIST,
-    TERRITORY_FIPS_LIST,
-    US_ABBR,
-    US_FIPS,
-    US_NAME,
-    TERRITORY_POSTALS,
-    CURRENT,
-    HISTORICAL,
-)
-from typing import cast, List
-from ingestion.het_types import GEO_TYPE, SEX_RACE_AGE_TYPE
-
-RACE = cast(SEX_RACE_AGE_TYPE, "race")
-
-
-FIPS_TO_STATE_TABLE_MAP = {
-    "01": "128",
-    "02": "2312",
-    "04": "764",
-    "05": "775",
-    "06": "781",
-    "08": "787",
-    "09": "793",
-    "10": "799",
-    "12": "167",
-    "13": "805",
-    "15": "173",
-    "16": "648",
-    "17": "649",
-    "18": "650",
-    "19": "651",
-    "20": "836",
-    "21": "653",
-    "22": "654",
-    "23": "655",
-    "24": "857",
-    "25": "657",
-    "26": "658",
-    "27": "879",
-    "28": "660",
-    "29": "661",
-    "30": "662",
-    "31": "276",
-    "32": "663",
-    "33": "664",
-    "34": "665",
-    "35": "209",
-    "36": "918",
-    "37": "667",
-    "38": "668",
-    "39": "669",
-    "40": "672",
-    "41": "673",
-    "42": "679",
-    "44": "685",
-    "45": "691",
-    "46": "697",
-    "47": "703",
-    "48": "710",
-    "49": "716",
-    "50": "722",
-    "51": "233",
-    "53": "739",
-    "54": "740",
-    "55": "746",
-    "56": "752",
-}
-
-# time_periods for entire dataset
-DEFAULT_CONGRESS_FIRST_YR = 1915
-DEFAULT_STLEG_FIRST_YR = 1983
-# TODO should be calc from most recent year in numerator source data. see GitHub #2897
-DEFAULT_LAST_YR = 2025
-
-# data urls
-US_CONGRESS_CURRENT_URL = "https://theunitedstates.io/congress-legislators/legislators-current.json"
-US_CONGRESS_HISTORICAL_URL = "https://theunitedstates.io/congress-legislators/legislators-historical.json"
-CAWP_LINE_ITEMS_FILE = "cawp-by_race_and_ethnicity_time_series.csv"
-
-
-CAWP_POP_PCT_COL = "cawp_population_pct"
-CAWP_POP_COL = "cawp_population"
-
-
-def get_stleg_url(id: str):
-    """
-    Constructs the needed url for each state's table that contains the
-    total number of state legislators per year. These tables are on the
-    state info pages, for example:
-    https://cawp.rutgers.edu/facts/state-state-information/alabama
-    """
-    return f"https://cawp.rutgers.edu/tablefield/export/paragraph/{id}/field_table/und/0"
-
-
-CAWP_MULTI = "Multiracial Alone"
-
-# CAWP labels
-CAWP_RACE_GROUPS_TO_STANDARD = {
-    "Asian American/Pacific Islander": Race.ASIAN_PAC.value,
-    "Latina": Race.HISP.value,
-    "Middle Eastern/North African": Race.MENA.value,
-    "Native American/Alaska Native/Native Hawaiian": Race.AIANNH.value,
-    "Black": Race.BLACK.value,
-    "White": Race.WHITE.value,
-    "Unavailable": Race.UNKNOWN.value,
-    "Other": Race.OTHER_STANDARD.value,
-    # will combine CAWP's "Multiracial Alone" with women who selected more than one specific race
-    CAWP_MULTI: Race.MULTI.value,
-}
-
-
-AIAN_API_RACES = [
-    "Asian American/Pacific Islander",
-    "Native American/Alaska Native/Native Hawaiian",
-]
-
-
-RACE_ETH = "race_ethnicity"
-NAME = "name"
-FIRST = "first"
-LAST = "last"
-TYPE = "type"
-ID = "id"
-STATE = "state"
-TERMS = "terms"
-START = "start"
-END = "end"
-FIRST_NAME = "first_name"
-LAST_NAME = "last_name"
-POSITION = "position"
-LEVEL = "level"
-YEAR = "year"
-
-CONGRESS = "Congress"
-STATE_LEG = "State Legislative"
-TERRITORY_LEG = "Territorial/D.C."
-
-STATE_COLS = [std_col.STATE_FIPS_COL, std_col.STATE_POSTAL_COL, std_col.STATE_NAME_COL]
-
-MERGE_COLS = [std_col.TIME_PERIOD_COL, *STATE_COLS, RACE_ETH]
-
-POSITION_LABELS = {
-    CONGRESS: {
-        "U.S. Representative": "U.S. Rep.",
-        "rep": "U.S. Rep.",
-        "sen": "U.S. Sen.",
-        "U.S. Senator": "U.S. Sen.",
-        "U.S. Delegate": "U.S. Del.",
-    },
-    STATE_LEG: {
-        "State Representative": "State Rep.",
-        "State Senator": "State Sen.",
-        "Territorial/D.C. Representative": "Rep.",
-        "Territorial/D.C. Senator": "Sen.",
-    },
-}
-
-
-class CAWPTimeData(DataSource):
-    @staticmethod
-    def get_id():
-        return "CAWP_TIME_DATA"
-
-    @staticmethod
-    def get_table_name():
-        return "cawp_time_data"
-
-    def upload_to_gcs(self, _, **attrs):
-        raise NotImplementedError("upload_to_gcs should not be called for CAWPTimeData")
-
-    def write_to_bq(self, dataset, gcs_bucket, **attrs):
-        base_df = self.generate_base_df()
-        df_names = base_df.copy()
-        df_names = self.generate_names_breakdown(df_names)
-        column_types = gcs_to_bq_util.get_bq_column_types(df_names, [])
-        # adds a df of legislator names for user download
-        gcs_to_bq_util.add_df_to_bq(
-            df_names,
-            dataset,
-            "race_and_ethnicity_state_historical_names",
-            column_types=column_types,
-        )
-
-        for geo_level in [STATE_LEVEL, NATIONAL_LEVEL]:
-            df = base_df.copy()
-            df = self.generate_breakdown(df, geo_level)
-
-            # make _historical table with limited cols
-            df_historical = df.copy()
-            df_historical = df.drop(
-                columns=[
-                    std_col.CONGRESS_COUNT,
-                    std_col.W_THIS_RACE_CONGRESS_COUNT,
-                    std_col.PCT_OF_W_CONGRESS,
-                    std_col.STLEG_COUNT,
-                    std_col.W_THIS_RACE_STLEG_COUNT,
-                    std_col.PCT_OF_W_STLEG,
-                    CAWP_POP_COL,
-                    CAWP_POP_PCT_COL,
-                ]
-            )
-            float_cols = [
-                std_col.PCT_OF_CONGRESS,
-                std_col.W_CONGRESS_PCT_INEQUITY,
-                std_col.PCT_OF_STLEG,
-                std_col.W_STLEG_PCT_INEQUITY,
-            ]
-
-            column_types = gcs_to_bq_util.get_bq_column_types(df_historical, float_cols)
-            df_historical = df_historical.drop_duplicates()
-            gcs_to_bq_util.add_df_to_bq(
-                df_historical,
-                dataset,
-                gcs_to_bq_util.make_bq_table_id(std_col.RACE_OR_HISPANIC_COL, geo_level, HISTORICAL),
-                column_types=column_types,
-            )
-
-            # make _current table with limited rows
-            df_current = df.copy()
-            df_current = df_current.drop(
-                columns=[
-                    std_col.W_CONGRESS_PCT_INEQUITY,
-                    std_col.W_STLEG_PCT_INEQUITY,
-                ]
-            )
-            # drop unneeded rows from current
-            df_current = preserve_only_current_time_period_rows(df_current)
-
-            float_cols = [
-                std_col.CONGRESS_COUNT,
-                std_col.W_THIS_RACE_CONGRESS_COUNT,
-                std_col.PCT_OF_CONGRESS,
-                std_col.PCT_OF_W_CONGRESS,
-                std_col.STLEG_COUNT,
-                std_col.W_THIS_RACE_STLEG_COUNT,
-                std_col.PCT_OF_STLEG,
-                std_col.PCT_OF_W_STLEG,
-                CAWP_POP_COL,
-                CAWP_POP_PCT_COL,
-            ]
-
-            column_types = gcs_to_bq_util.get_bq_column_types(df_current, float_cols)
-            df_current = df_current.drop_duplicates()
-            gcs_to_bq_util.add_df_to_bq(
-                df_current,
-                dataset,
-                gcs_to_bq_util.make_bq_table_id(std_col.RACE_OR_HISPANIC_COL, geo_level, CURRENT),
-                column_types=column_types,
-            )
-
-    # CLASS METHODS
-
-    def generate_base_df(self):
-        """Creates a dataframe with the raw counts by state by year by race of:
-        all congress members, all women congress members,
-        and women congress members of the row's race"""
-
-        # fetch and form data
-        women_us_congress_df, women_state_leg_df = get_women_dfs()
-        us_congress_totals_df = get_us_congress_totals_df()
-        state_leg_totals_df = get_state_leg_totals_df()
-
-        # create ROWS for the "All" race
-        df_alls_rows = build_base_rows_df(
-            us_congress_totals_df,
-            state_leg_totals_df,
-            women_us_congress_df,
-            women_state_leg_df,
-            [Race.ALL.value],
-        )
-
-        # create ROWS for each CAWP race group
-        df_by_races_rows = build_base_rows_df(
-            us_congress_totals_df,
-            state_leg_totals_df,
-            women_us_congress_df,
-            women_state_leg_df,
-            list(CAWP_RACE_GROUPS_TO_STANDARD.keys()),
-        )
-
-        # append ROWS for combo race AIAN_API
-        df_by_races_rows = add_aian_api_rows(df_by_races_rows)
-
-        # combine ROWS together from ALLS ROWS and BY RACES rows
-        df = pd.concat([df_alls_rows, df_by_races_rows])
-
-        # standardize race labels
-        df[std_col.RACE_CATEGORY_ID_COL] = df[RACE_ETH].apply(
-            lambda x: ("ALL" if x == Race.ALL.value else CAWP_RACE_GROUPS_TO_STANDARD.get(x, x))
-        )
-        std_col.add_race_columns_from_category_id(df)
-        df = df.drop(columns=[RACE_ETH])
-
-        df = df.sort_values(by=[std_col.TIME_PERIOD_COL, *STATE_COLS, std_col.RACE_OR_HISPANIC_COL]).reset_index(
-            drop=True
-        )
-
-        df = df.drop(
-            [
-                std_col.W_ALL_RACES_CONGRESS_NAMES,
-                std_col.W_ALL_RACES_STLEG_NAMES,
-            ],
-            axis=1,
-        )
-
-        names_cols = [
-            std_col.CONGRESS_NAMES,
-            std_col.W_THIS_RACE_CONGRESS_NAMES,
-            std_col.W_THIS_RACE_STLEG_NAMES,
-        ]
-
-        for col in names_cols:
-            # Replace nulls with empty lists
-            df.loc[df[col].isnull(), col] = df.loc[df[col].isnull(), col].apply(lambda x: [])
-            # Convert lists to comma-separated strings
-            df[col] = df[col].apply(lambda item: ",".join(map(str, item)))
-
-        # remove brackets and inner quotes, leaving just comma separated names
-        df[names_cols] = df[names_cols].replace(["'", "[", "]"], "")
-
-        # Set CONGRESS_NAMES to None where RACE_CATEGORY_ID_COL is not ALL
-        df.loc[df[std_col.RACE_CATEGORY_ID_COL] != Race.ALL.value, std_col.CONGRESS_NAMES] = None
-
-        return df
-
-    def generate_breakdown(self, _df, geo_level: str):
-        """Takes df with rows per year/race incl ALL/state and calculates the metrics
-        shown on the frontend
-
-        Parameters:
-            df: with columns for state info, CAWP "race_ethnicity",
-                "time_period" years, along with the raw counts for total,
-                all women, and women of each race
-            geo_level:
-                "national" or "state"
-        Returns [df, bq_table_name]:
-            df: with calculated columns for share of congress,
-                share of women in congress, percent relative inequity
-            bq_table_name: string name used for writing each breakdown to bq"""
-
-        df = _df.copy()
-
-        # we don't need the names columns anymore
-        df = df.drop(
-            [
-                std_col.CONGRESS_NAMES,
-                std_col.W_THIS_RACE_CONGRESS_NAMES,
-                std_col.W_THIS_RACE_STLEG_NAMES,
-            ],
-            axis=1,
-        )
-
-        if geo_level == NATIONAL_LEVEL:
-            df = combine_states_to_national(df)
-
-        # calculate rates of representation
-        df[std_col.PCT_OF_CONGRESS] = round(
-            df[std_col.W_THIS_RACE_CONGRESS_COUNT] / df[std_col.CONGRESS_COUNT] * 100, 1
-        )
-        df[std_col.PCT_OF_W_CONGRESS] = round(
-            df[std_col.W_THIS_RACE_CONGRESS_COUNT] / df[std_col.W_ALL_RACES_CONGRESS_COUNT] * 100,
-            1,
-        ).fillna(0)
-        df[std_col.PCT_OF_STLEG] = round(df[std_col.W_THIS_RACE_STLEG_COUNT] / df[std_col.STLEG_COUNT] * 100, 1)
-        df[std_col.PCT_OF_W_STLEG] = round(
-            df[std_col.W_THIS_RACE_STLEG_COUNT] / df[std_col.W_ALL_RACES_STLEG_COUNT] * 100,
-            1,
-        )
-
-        # drop the ALL WOMEN counts since that data is available on the "All" race rows
-        df = df.drop(
-            columns=[
-                std_col.W_ALL_RACES_CONGRESS_COUNT,
-                std_col.W_ALL_RACES_STLEG_COUNT,
-            ]
-        )
-
-        df = merge_utils.merge_yearly_pop_numbers(df, RACE, cast(GEO_TYPE, geo_level))
-
-        df = generate_pct_rel_inequity_col(
-            df,
-            std_col.PCT_OF_W_CONGRESS,
-            std_col.POPULATION_PCT_COL,
-            std_col.W_CONGRESS_PCT_INEQUITY,
-        )
-        df = zero_out_pct_rel_inequity(
-            df,
-            cast(GEO_TYPE, geo_level),
-            RACE,
-            {std_col.PCT_OF_CONGRESS: std_col.W_CONGRESS_PCT_INEQUITY},
-            std_col.POPULATION_PCT_COL,
-        )
-
-        df = generate_pct_rel_inequity_col(
-            df,
-            std_col.PCT_OF_W_STLEG,
-            std_col.POPULATION_PCT_COL,
-            std_col.W_STLEG_PCT_INEQUITY,
-        )
-        df = zero_out_pct_rel_inequity(
-            df,
-            cast(GEO_TYPE, geo_level),
-            RACE,
-            {std_col.PCT_OF_STLEG: std_col.W_STLEG_PCT_INEQUITY},
-            std_col.POPULATION_PCT_COL,
-        )
-
-        sort_cols = [std_col.TIME_PERIOD_COL, *STATE_COLS, std_col.RACE_CATEGORY_ID_COL]
-
-        df = df.sort_values(by=sort_cols).reset_index(drop=True)
-
-        # don't keep pct_rates for AIAN_API
-        mask = df[std_col.RACE_CATEGORY_ID_COL] == Race.AIAN_API.value
-        df.loc[mask, std_col.PCT_OF_CONGRESS] = None
-        df.loc[mask, std_col.PCT_OF_STLEG] = None
-
-        df = df.rename(
-            columns={
-                std_col.POPULATION_COL: CAWP_POP_COL,
-                std_col.POPULATION_PCT_COL: CAWP_POP_PCT_COL,
-            }
-        )
-
-        return df
-
-    def generate_names_breakdown(self, df: pd.DataFrame):
-        """Create the state-level, names-only df which will be available for download
-        but not displayed on the tracker
-
-        Parameters:
-            df: The generate_base_df created previously that contains columns
-                for both _counts and _names along with year and state info
-        Returns:
-            the same df with the _counts removed, only the _names, state
-                and year cols remaining"""
-
-        df = df[
-            [
-                std_col.TIME_PERIOD_COL,
-                std_col.STATE_FIPS_COL,
-                std_col.STATE_NAME_COL,
-                std_col.RACE_CATEGORY_ID_COL,
-                std_col.RACE_OR_HISPANIC_COL,
-                std_col.CONGRESS_NAMES,
-                std_col.W_THIS_RACE_CONGRESS_NAMES,
-                std_col.W_THIS_RACE_STLEG_NAMES,
-            ]
-        ]
-
-        return df
-
-
-# HELPER FUNCTIONS
-
-
-def scaffold_df_by_year_by_state_by_race_list(race_list: List[str], first_year: int):
-    """Creates the scaffold df with a row for every STATE/YEAR/RACE_ETH IN race_list combo
-    Parameters:
-        race_list: list of strings to serve as values in the "race_ethnicity" column
-        first_year: int year to start building the scaffold e.g. 1983
-    Returns:
-        df with a row for every combo of `race_list` race, years, and state/territories
-        including columns for "state_name", "state_postal" and "state_fips" """
-    # start with single column of all state-level fips as our df template
-    fips_list = get_state_level_fips()
-    df = pd.DataFrame(
-        {
-            std_col.STATE_FIPS_COL: [*fips_list],
-        }
-    )
-
-    # explode to every combo of state/year
-    years = get_consecutive_time_periods(first_year=first_year)
-    df[std_col.TIME_PERIOD_COL] = [years] * len(df)
-    df = df.explode(std_col.TIME_PERIOD_COL).reset_index(drop=True)
-
-    # merge in FIPS codes to the scaffold df
-    df = merge_utils.merge_state_ids(df, keep_postal=True)
-    df[RACE_ETH] = [race_list] * len(df)
-    df = df.explode(RACE_ETH)
-
-    return df
-
-
-def get_us_congress_totals_df():
-    """Fetches historic and current congress data, combines them, and iterates over
-    each Congress member and their terms served to generate a dataframe.
-
-    Returns:
-        df with rows per legislator-term and
-        columns "time_period" by year and "state_postal" """
-
-    # load US congress data for total_counts
-    raw_historical_congress_json = gcs_to_bq_util.fetch_json_from_web(US_CONGRESS_HISTORICAL_URL)
-    raw_current_congress_json = gcs_to_bq_util.fetch_json_from_web(US_CONGRESS_CURRENT_URL)
-
-    raw_legislators_json = [*raw_historical_congress_json, *raw_current_congress_json]
-
-    us_congress_totals_list_of_dict = []
-    years = get_consecutive_time_periods()
-
-    # iterate through each legislator
-    for legislator in raw_legislators_json:
-        # and each term they served
-        for term in legislator[TERMS]:
-            term_years = list(range(int(term[START][:4]), int(term[END][:4]) + 1))
-
-            # and each year of each term
-            for year in term_years:
-                year = str(year)
-                title = (
-                    f"{POSITION_LABELS[CONGRESS][term[TYPE]]}" if term[STATE] not in TERRITORY_POSTALS else "U.S. Del."
-                )
-                full_name = f"{title} {legislator[NAME][FIRST]} {legislator[NAME][LAST]}"
-                entry = {
-                    ID: legislator[ID]["govtrack"],
-                    NAME: full_name,
-                    TYPE: term[TYPE],
-                    std_col.STATE_POSTAL_COL: term[STATE],
-                    std_col.TIME_PERIOD_COL: year,
-                }
-                # add entry of service for id/year/state.
-                # avoid double counting, CAWP only has 1 entry per leg. per year
-                if year in years and entry not in us_congress_totals_list_of_dict:
-                    us_congress_totals_list_of_dict.append(entry)
-
-    # convert to df
-    df = pd.DataFrame.from_dict(us_congress_totals_list_of_dict)
-
-    # get names of all TOTAL members in lists per row
-    df = df.groupby([std_col.STATE_POSTAL_COL, std_col.TIME_PERIOD_COL])[NAME].apply(list).reset_index()
-    df = df.rename(columns={NAME: std_col.CONGRESS_NAMES})
-
-    # get counts of all TOTAL members in lists per row
-    df[std_col.CONGRESS_COUNT] = df[std_col.CONGRESS_NAMES].apply(lambda list: len(list)).astype(float)
-
-    return df
-
-
-def merge_total_cols(scaffold_congress_df, scaffold_stleg_df, us_congress_df, state_leg_df):
-    """Merges previously made congress df and state_leg_df info into the incoming scaffold df
-    Parameters:
-        scaffold_congress_df: df containing a row for every combo of
-            "time_period" for valid us congress totals X "state_postal" X "race_ethnicity
-        scaffold_stleg_df: df containing a row for every combo of
-            "time_period" for valid state leg totals X "state_postal" X "race_ethnicity
-        congress_df: df containing a row for every legislator-term
-        state_leg_df: df containing a row for every "time_period" X "state_fips"
-
-    Returns:
-        df with a column "us_congress_total_count" ints count of total members in the state/year,
-        and column "us_congress_total_names" a string list of those same members"""
-
-    # merge in CONGRESS calculated counts and name lists by state/year where they exist;
-    df_congress_totals = pd.merge(
-        scaffold_congress_df,
-        us_congress_df,
-        on=[std_col.TIME_PERIOD_COL, std_col.STATE_POSTAL_COL],
-        how="left",
-    )
-    # fill counts with 0 and names with empty string where no info available
-    df_congress_totals[std_col.CONGRESS_COUNT] = df_congress_totals[std_col.CONGRESS_COUNT].fillna(0)
-    df_congress_totals[std_col.CONGRESS_NAMES] = df_congress_totals[std_col.CONGRESS_NAMES].fillna("")
-
-    # merge in STATE LEG counts by state/year where they exist;
-    df_stateleg_totals = pd.merge(
-        scaffold_stleg_df,
-        state_leg_df,
-        on=[std_col.TIME_PERIOD_COL, std_col.STATE_FIPS_COL],
-        how="left",
-    )
-
-    df = pd.merge(df_congress_totals, df_stateleg_totals, on=MERGE_COLS, how="left")
-
-    # fill counts with null where no info available
-    df[std_col.STLEG_COUNT] = df[std_col.STLEG_COUNT].astype(float)
-
-    return df
-
-
-def get_women_dfs():
-    """Fetches CAWP data counts of women by race by year by state,
-    generates two dfs, each with rows for every woman
-    in U.S. Congress or state/territory legislature in any year
-
-    Returns :
-        [women_us_congress_df, women_state_leg_df] each with rows per woman legislator
-            columns "time_period" by year and "state_postal", "race_ethnicity"
-            with specific CAWP race strings"""
-
-    df = gcs_to_bq_util.load_csv_as_df_from_data_dir("cawp_time", CAWP_LINE_ITEMS_FILE)
-
-    # keep only needed cols
-    df = df[[ID, YEAR, STATE, FIRST_NAME, LAST_NAME, POSITION, RACE_ETH]]
-
-    # standardize CAWP state names as postal
-    df[std_col.STATE_POSTAL_COL] = df[STATE].apply(get_postal_from_cawp_phrase)
-
-    # merge in FIPS codes
-    df = merge_utils.merge_state_ids(df, keep_postal=True)
-    df = df.drop(columns=[STATE])
-
-    # rename year
-    df = df.rename(columns={YEAR: std_col.TIME_PERIOD_COL})
-    df[std_col.TIME_PERIOD_COL] = df[std_col.TIME_PERIOD_COL].astype(str)
-
-    women_dfs = []
-
-    for gov_level in [CONGRESS, STATE_LEG]:
-        # remove non-legislative line items
-        df_gov_level = df.copy().loc[df[POSITION].isin(POSITION_LABELS[gov_level].keys())]
-        # standardize gov. titles between sources
-        df_gov_level[POSITION] = df_gov_level[POSITION].apply(lambda x: POSITION_LABELS[gov_level][x])
-        # consolidate name columns
-        df_gov_level[NAME] = df_gov_level[POSITION] + " " + df_gov_level[FIRST_NAME] + " " + df_gov_level[LAST_NAME]
-        df_gov_level = df_gov_level.drop(columns=[FIRST_NAME, LAST_NAME, POSITION])
-        women_dfs.append(df_gov_level)
-    return women_dfs
-
-
-def merge_women_cols(scaffold_df, women_df, gov_level: str, preserve_races: bool = False):
-    """Merges previously made CAWP df info into the incoming scaffold df
-    Parameters:
-        scaffold_df: df containing a row for every combo of
-            "time_period" X "state_postal" X "race_ethnicity
-        women_df: df containing a row for every woman in US Congress or State Leg
-        gov_level: string to choose the level of government to add cols for,
-            either `state_leg` or `us_congress`
-        preserve_races (optional boolean): if True will calculate the counts and names
-            per race and merge "_this_race" cols, if False will perform the
-            calculations for the Race.ALL.value race group and merge the "_all_races" cols
-
-    Returns:
-        df with rows for "time_period" X "state_postal" X (optional RACE_ETH), and
-        _count cols for state_leg and us_congress of total women members in
-            the state/year/each race or all races
-        and _names cols with string lists of those same members"""
-
-    df = women_df.copy()
-
-    groupby_cols = [*STATE_COLS, std_col.TIME_PERIOD_COL]
-
-    needed_cols = groupby_cols[:]
-    needed_cols.append(NAME)
-
-    names_col = None
-    count_col = None
-
-    if preserve_races:
-        df = handle_other_and_multi_races(df)
-        needed_cols.append(RACE_ETH)
-        groupby_cols.append(RACE_ETH)
-
-        if gov_level == CONGRESS:
-            count_col = std_col.W_THIS_RACE_CONGRESS_COUNT
-            names_col = std_col.W_THIS_RACE_CONGRESS_NAMES
-        elif gov_level == STATE_LEG:
-            count_col = std_col.W_THIS_RACE_STLEG_COUNT
-            names_col = std_col.W_THIS_RACE_STLEG_NAMES
-    else:
-        if gov_level == CONGRESS:
-            count_col = std_col.W_ALL_RACES_CONGRESS_COUNT
-            names_col = std_col.W_ALL_RACES_CONGRESS_NAMES
-        elif gov_level == STATE_LEG:
-            count_col = std_col.W_ALL_RACES_STLEG_COUNT
-            names_col = std_col.W_ALL_RACES_STLEG_NAMES
-
-    # remove unneeded cols
-    df = df[needed_cols]
-
-    # collapse yr/state/race rows, combine names into a list
-    df = df.groupby(groupby_cols)[NAME].apply(list).reset_index()
-    df = df.rename(columns={NAME: names_col})
-
-    # generate counts by counting the # of names
-    df[count_col] = df[names_col].apply(lambda list: len(list)).astype(float)
-
-    df = pd.merge(scaffold_df, df, on=groupby_cols, how="left")
-    df[count_col] = df[count_col].fillna(0)
-    df[names_col] = df[names_col].fillna("")
-
-    return df
-
-
-def get_state_leg_totals_df():
-    """Fetches each individual CAWP state info page's state legislature
-    table, combines into a single cleaned df. Nulls everything before 1983;
-    CAWPs totals are problematic between 1975-1982 and missing before that.
-
-    Returns: df with "time_period", "state_fips", and stleg
-        total_ and total_women cols
-
-    """
-
-    territory_dfs = []
-    for fips in TERRITORY_FIPS_LIST:
-        filename = f"cawp_state_leg_{fips}.csv"
-        territory_df = gcs_to_bq_util.load_csv_as_df_from_data_dir(
-            "cawp_time", filename, dtype={"state_fips": str, "time_period": str}
-        )
-        territory_dfs.append(territory_df)
-    df_rows_by_territory = pd.concat(territory_dfs)
-
-    state_dfs = []
-    for fips, id in FIPS_TO_STATE_TABLE_MAP.items():
-        state_df = gcs_to_bq_util.load_csv_as_df_from_web(get_stleg_url(id), dtype=str)
-
-        # remove weird chars from col headers
-        state_df.columns = state_df.columns.str.replace(r"\W", "", regex=True)
-
-        # standardize the year col
-        state_df = state_df.rename(columns={"Year": std_col.TIME_PERIOD_COL})
-        # Drop rows where year is NaN
-        state_df = state_df.dropna(subset=[std_col.TIME_PERIOD_COL])
-
-        state_df[std_col.TIME_PERIOD_COL] = state_df[std_col.TIME_PERIOD_COL].astype(str)
-
-        # extract totals
-        state_df[[std_col.W_ALL_RACES_STLEG_COUNT, std_col.STLEG_COUNT]] = state_df[
-            "TotalWomenTotalLegislature"
-        ].str.split("/", n=1, expand=True)
-
-        # keep only needed cols
-        state_df = state_df[[std_col.TIME_PERIOD_COL, std_col.STLEG_COUNT]]
-
-        # append this state df to the list
-        state_df[std_col.STATE_FIPS_COL] = fips
-        state_dfs.append(state_df)
-
-    # combine all state ROWS into one big df
-    df_rows_by_state = pd.concat(state_dfs)
-
-    # combine all territory ROWS as well
-    df = pd.concat([df_rows_by_state, df_rows_by_territory])
-    df = df.sort_values(by=[std_col.TIME_PERIOD_COL, std_col.STATE_FIPS_COL]).reset_index(drop=True)
-
-    return df
-
-
-def combine_states_to_national(df):
-    """Takes the df that contains rows for every year/race by state and territory,
-    and combines those rows into a national dataset
-
-    Parameters:
-        df: dataframe containing a row for every combination of state/race/year
-    Output:
-        df same dataframe summed to a national level with a row per race/year"""
-
-    state_cols = [*STATE_COLS]
-    groupby_cols = [
-        std_col.TIME_PERIOD_COL,
-        std_col.RACE_CATEGORY_ID_COL,
-        std_col.RACE_OR_HISPANIC_COL,
-    ]
-    df_counts = df.copy().drop(state_cols, axis=1)
-    df_counts = df_counts.groupby(groupby_cols, as_index=False)[
-        [
-            std_col.CONGRESS_COUNT,
-            std_col.W_ALL_RACES_CONGRESS_COUNT,
-            std_col.W_THIS_RACE_CONGRESS_COUNT,
-            std_col.STLEG_COUNT,
-            std_col.W_ALL_RACES_STLEG_COUNT,
-            std_col.W_THIS_RACE_STLEG_COUNT,
-        ]
-    ].agg("sum", min_count=1)
-
-    df_counts[std_col.STATE_FIPS_COL] = US_FIPS
-    df_counts[std_col.STATE_NAME_COL] = US_NAME
-    df_counts[std_col.STATE_POSTAL_COL] = US_ABBR
-
-    return df_counts
-
-
-def get_postal_from_cawp_phrase(cawp_place_phrase: str):
-    """Swap CAWP place phrase found in the LINE ITEM table
-    `{STATE_COL_LINE NAME} - {CODE}` with the standard 2 letter code
-
-    Parameters:
-        cawp_place_phrase: str
-    Returns:
-        string of standard 2-letter postal code"""
-
-    # swap out non-standard 2 letter codes
-    cawp_place_phrase = {
-        "American Samoa - AM": "American Samoa - AS",
-        "Northern Mariana Islands - MI": "Northern Mariana Islands - MP",
-    }.get(cawp_place_phrase, cawp_place_phrase)
-
-    return cawp_place_phrase.split(" - ")[1]
-
-
-def get_consecutive_time_periods(first_year: int = DEFAULT_CONGRESS_FIRST_YR, last_year: int = DEFAULT_LAST_YR):
-    """Generates a list of consecutive time periods in the "YYYY" format
-
-    Parameters:
-        first_year: optional int to start the list; defaults to 1915
-            which is two years before the first woman in US Congress
-        last_year: optional int to be the last element in the list
-            other than the default
-    Returns:
-        a list of string years (e.g. ["1999", "2000", "2001"])"""
-    return [str(x) for x in list(range(first_year, last_year + 1))]
-
-
-def get_state_level_fips():
-    """Returns list of 2-letter strings for state and territory fips codes"""
-    return STATE_LEVEL_FIPS_LIST
-
-
-def add_aian_api_rows(df):
-    """Adds new rows for the combined AIAN_API race group"""
-
-    # only keep rows with years that will get population
-    target_time_periods = get_consecutive_time_periods(
-        first_year=int(ACS_EARLIEST_YEAR), last_year=int(ACS_CURRENT_YEAR)
-    )
-    df_aian_api_rows = df[df[std_col.TIME_PERIOD_COL].isin(target_time_periods)]
-
-    # only keep rows with races to be combined
-    df_aian_api_rows = df_aian_api_rows.loc[df_aian_api_rows[RACE_ETH].isin(AIAN_API_RACES)]
-
-    level_names_col_map = {
-        CONGRESS: std_col.W_THIS_RACE_CONGRESS_NAMES,
-        STATE_LEG: std_col.W_THIS_RACE_STLEG_NAMES,
-    }
-    level_count_col_map = {
-        CONGRESS: std_col.W_THIS_RACE_CONGRESS_COUNT,
-        STATE_LEG: std_col.W_THIS_RACE_STLEG_COUNT,
-    }
-    level_denom_cols_map = {
-        CONGRESS: [
-            std_col.CONGRESS_COUNT,
-            std_col.CONGRESS_NAMES,
-            std_col.W_ALL_RACES_CONGRESS_COUNT,
-            std_col.W_ALL_RACES_CONGRESS_NAMES,
-        ],
-        STATE_LEG: [
-            std_col.STLEG_COUNT,
-            std_col.W_ALL_RACES_STLEG_COUNT,
-            std_col.W_ALL_RACES_STLEG_NAMES,
-        ],
-    }
-
-    aian_api_dfs = []
-
-    for gov_level in [CONGRESS, STATE_LEG]:
-        # specific columns needed for this level of government
-        names_col = level_names_col_map[gov_level]
-        count_col = level_count_col_map[gov_level]
-        level_denom_cols = level_denom_cols_map[gov_level]
-
-        # only keep needed columns
-        df_aian_api_rows_gov_level = df_aian_api_rows.copy()[
-            [std_col.TIME_PERIOD_COL, *STATE_COLS, names_col]
-        ].reset_index(drop=True)
-
-        # combine the race rows, and their lists of names
-        df_aian_api_rows_gov_level = df_aian_api_rows_gov_level.groupby(
-            [std_col.TIME_PERIOD_COL, *STATE_COLS], as_index=False
-        )[names_col].agg(lambda nested_list: [x for list in nested_list for x in list])
-
-        # remove any duplicates if a women was in both of the combined race groups
-        df_aian_api_rows_gov_level[names_col] = df_aian_api_rows_gov_level[names_col].apply(set).apply(list)
-
-        df_aian_api_rows_gov_level[count_col] = (
-            df_aian_api_rows_gov_level[names_col].apply(lambda list: len(list)).astype(float)
-        )
-        df_aian_api_rows_gov_level[RACE_ETH] = Race.AIAN_API.value
-        df_aian_api_rows_gov_level = df_aian_api_rows_gov_level.reset_index(drop=True)
-
-        # re-merge with this to preserve the non-summed rows like "total_congress_count", etc
-        # could use either Asian or AIAN, the totals would be the same
-        orig_df = df.copy()
-        df_denom_cols_aian_api_rows = orig_df.copy().loc[orig_df[RACE_ETH] == "Asian American/Pacific Islander"]
-
-        denom_cols = [std_col.TIME_PERIOD_COL, *STATE_COLS, *level_denom_cols]
-
-        df_denom_cols = df_denom_cols_aian_api_rows[denom_cols].reset_index(drop=True)
-
-        # add back on the COLUMNS that didn't need to sum
-        df_aian_api_rows_gov_level = pd.merge(
-            df_aian_api_rows_gov_level,
-            df_denom_cols,
-            on=[std_col.TIME_PERIOD_COL, *STATE_COLS],
-        ).reset_index(drop=True)
-
-        # store for later merging on cols
-        aian_api_dfs.append(df_aian_api_rows_gov_level)
-
-    # merge combo race rows state_leg cols and us_congress cols
-    aian_api_rows_state_leg_cols_df, aian_api_rows_us_congress_cols_df = aian_api_dfs
-
-    df_aian_api_rows = pd.merge(
-        aian_api_rows_state_leg_cols_df,
-        aian_api_rows_us_congress_cols_df,
-        on=[std_col.TIME_PERIOD_COL, *STATE_COLS, RACE_ETH],
-    ).reset_index(drop=True)
-
-    # add COMBO AIAN_API RACE ROWS onto the original race groups ROWS
-    df = pd.concat([df, df_aian_api_rows], axis="rows").reset_index(drop=True)
-
-    return df
-
-
-def build_base_rows_df(
-    us_congress_totals_df,
-    state_leg_totals_df,
-    women_us_congress_df,
-    women_state_leg_df,
-    race_list: List[str],
-):
-    """Builds out a scaffold of rows with YEAR/STATE/RACE combos,
-    then merges columns for:
-    - TOTAL CONGRESS, WOMEN IN CONGRESS, WOMEN THIS RACE IN CONGRESS
-    - TOTAL STATE LEG, WOMEN IN STATE LEG, WOMEN THIS RACE STATE LEG
-
-    Parameters:
-        us_congress_totals_df: prev. processed df with congress info from unitedstates project
-        state_leg_totals_df: prev. processed df with state leg info from CAWP state info pages
-        women_us_congress_df: prev. processed df with women in US Congress info from CAWP database
-        women_state_leg_df: prev. processed df with women in state leg. info from CAWP database
-        race_list: a list of strings representing which races should be included in this base chunk
-
-    Returns: a df with rows per year/state/race from race list, with columns incl.
-        US CONGRESS and STATE LEG counts for TOTAL, WOMEN ALL RACE, and WOMEN THIS RACE
-    """
-
-    # create chunks with needed COLUMNS
-    df_congress_scaffold = scaffold_df_by_year_by_state_by_race_list(race_list, DEFAULT_CONGRESS_FIRST_YR)
-    df_stleg_scaffold = scaffold_df_by_year_by_state_by_race_list(race_list, DEFAULT_STLEG_FIRST_YR)
-
-    df_total_cols = merge_total_cols(
-        df_congress_scaffold.copy(),
-        df_stleg_scaffold,
-        us_congress_totals_df,
-        state_leg_totals_df,
-    )
-
-    df_w_any_race_us_congress_cols = merge_women_cols(df_congress_scaffold.copy(), women_us_congress_df, CONGRESS)
-
-    # intentionally merging onto congress scaffold because we can have
-    #  STLEG WOMEN COUNTS back further than the STLEG TOTALS
-    df_w_any_race_state_leg_cols = merge_women_cols(df_congress_scaffold.copy(), women_state_leg_df, STATE_LEG)
-
-    # for the ALL rows, the ALL_W cols will be the same as the W_THIS_RACE cols
-    # so don't need to waste time recalculating them
-    if race_list == [Race.ALL.value]:
-        df_w_this_race_us_congress_cols = df_w_any_race_us_congress_cols.copy().rename(
-            columns={
-                std_col.W_ALL_RACES_CONGRESS_NAMES: std_col.W_THIS_RACE_CONGRESS_NAMES,
-                std_col.W_ALL_RACES_CONGRESS_COUNT: std_col.W_THIS_RACE_CONGRESS_COUNT,
-            }
-        )
-        df_w_this_race_state_leg_cols = df_w_any_race_state_leg_cols.copy().rename(
-            columns={
-                std_col.W_ALL_RACES_STLEG_NAMES: std_col.W_THIS_RACE_STLEG_NAMES,
-                std_col.W_ALL_RACES_STLEG_COUNT: std_col.W_THIS_RACE_STLEG_COUNT,
-            }
-        )
-    else:
-        df_w_this_race_us_congress_cols = merge_women_cols(
-            df_congress_scaffold.copy(),
-            women_us_congress_df,
-            CONGRESS,
-            preserve_races=True,
-        )
-        df_w_this_race_state_leg_cols = merge_women_cols(
-            df_congress_scaffold.copy(),
-            women_state_leg_df,
-            STATE_LEG,
-            preserve_races=True,
-        )
-
-    # combine COLUMN chunks
-    df = pd.merge(df_congress_scaffold, df_total_cols, on=MERGE_COLS)
-    df = pd.merge(df, df_w_any_race_us_congress_cols, on=MERGE_COLS)
-    df = pd.merge(df, df_w_this_race_us_congress_cols, on=MERGE_COLS)
-    df = pd.merge(df, df_w_any_race_state_leg_cols, on=MERGE_COLS)
-    df = pd.merge(df, df_w_this_race_state_leg_cols, on=MERGE_COLS)
-
-    return df
-
-
-def handle_other_and_multi_races(df):
-    """
-    Parameters:
-        df which includes rows per women legislator, with a RACE_ETH column
-           containing a string of either: a single CAWP race or multiple
-           CAWP races separated by commas
-
-    Returns:
-        df where the original comma-containing "multiple specific race" rows have their
-        RACE_ETH value replaced by a list of specific CAWP races. Equivalent rows
-        are added with race labeled as MULTI_OTHER_TMP, and "Multiracial Alone"
-        are renamed, allowing these 2 types of multi- to be combined in the aggregation
-    """
-    # convert comma separated names string into list, doesn't affect single race strings
-    df[RACE_ETH] = df[RACE_ETH].str.split(", ")
-
-    # rows with multiple specific races will sum later with
-    # CAWP's incoming "multiracial alone"
-    df_multiple_specific = df[df[RACE_ETH].apply(lambda x: isinstance(x, str) and len(x) > 1)]
-    df_multiple_specific[RACE_ETH] = CAWP_MULTI
-    df = pd.concat([df, df_multiple_specific])
-
-    # create individual race rows from multiple specific rows
-    df = df.explode(RACE_ETH)
-
-    return df
