+++ python/datasources/data_sources.py
@@ -3,7 +3,7 @@ from datasources.acs_population import ACSPopulation
 from datasources.age_adjust_cdc_restricted import AgeAdjustCDCRestricted
 from datasources.age_adjust_cdc_hiv import AgeAdjustCDCHiv
 from datasources.bjs_incarceration import BJSIncarcerationData
-from datasources.cawp_time import CAWPTimeData
+from datasources.cawp import CAWPTimeData
 from datasources.cdc_hiv import CDCHIVData
 from datasources.cdc_restricted import CDCRestrictedData
 from datasources.cdc_vaccination_county import CDCVaccinationCounty
