+++ python/tests/data/cawp/golden_data/race_and_ethnicity_national_current.csv
@@ -1,0 +1,12 @@
+race_category_id,race_and_ethnicity,total_us_congress_count,women_this_race_us_congress_count,total_state_leg_count,women_this_race_state_leg_count,state_fips,state_name,state_postal,pct_share_of_us_congress,pct_share_of_women_us_congress,pct_share_of_state_leg,pct_share_of_women_state_leg,cawp_population,cawp_population_pct
+AIANNH,"Native American, Alaska Native, & Native Hawaiian",4.0,1.0,99.0,1.0,00,United States,US,25.0,33.3,1.0,5.0,,
+AIAN_API,"American Indian, Alaska Native, Asian & Pacific Islander",4.0,2.0,99.0,2.0,00,United States,US,,66.7,,10.0,22535913.0,6.7
+ALL,All,4.0,3.0,99.0,20.0,00,United States,US,75.0,100.0,20.2,100.0,334369975.0,100.0
+ASIAN_PAC,Asian American & Pacific Islander,4.0,1.0,99.0,1.0,00,United States,US,25.0,33.3,1.0,5.0,,
+BLACK,Black or African American,4.0,0.0,99.0,2.0,00,United States,US,0.0,0.0,2.0,10.0,41574960.0,12.4
+HISP,Hispanic or Latino,4.0,0.0,99.0,1.0,00,United States,US,0.0,0.0,1.0,5.0,64987193.0,19.4
+MENA,Middle Eastern & North African,4.0,0.0,99.0,0.0,00,United States,US,0.0,0.0,0.0,0.0,,
+MULTI,Two or more races,4.0,0.0,99.0,0.0,00,United States,US,0.0,0.0,0.0,0.0,29905388.0,8.9
+OTHER_STANDARD,Unrepresented race,4.0,0.0,99.0,0.0,00,United States,US,0.0,0.0,0.0,0.0,20804859.0,6.2
+UNKNOWN,Unknown race,4.0,0.0,99.0,0.0,00,United States,US,0.0,0.0,0.0,0.0,,
+WHITE,White,4.0,1.0,99.0,15.0,00,United States,US,25.0,33.3,15.2,75.0,219548855.0,65.7
