+++ python/tests/data/cawp/mock_cawp_state_leg_tables/cawp_state_leg_XX.csv
@@ -1,0 +1,2 @@
+-10,﻿Year,State Rank,Senate,Total Women/Total Senate,House,Total Women/Total House,Total Women/Total Legislature,% Total Women
+-9,2022,X,X, X/XX,X, X/XX, X/XX,X.X
