+++ python/tests/data/cawp/mock_territory_leg_tables/cawp_state_leg_ZZ_territory.csv
@@ -1,0 +1,2 @@
+time_period,state_fips,total_state_leg_count
+"1999","ZZ",12345
