--- python/tests/data/cawp_time/mock_cawp_state_leg_tables/cawp_state_leg_02.csv
@@ -1,51 +1,0 @@
--51,﻿Year,"State Rank",Senate,"Total Women/Total Senate",House,"Total Women/Total House","Total Women/Total Legislature","% Total Women"
--50,2024,25,"2D | 3R",5/20,"7D | 6R | 2I",15/40,20/60,33.3
--49,2023,24,"2D | 3R",5/20,"7D | 6R | 2I",15/40,20/60,33.3
--48,2022,27,"1D | 4R"," 5/20","7D | 6R"," 13/40"," 18/60",30.0
--47,2021,28,"1D | 4R"," 5/20","7D | 6R"," 13/40"," 18/60",30.0
--46,2020,10,"1D | 5R"," 6/20","6D | 10R"," 16/40"," 22/60",36.7
--45,2019,7,"1D | 5R"," 6/20","6D | 11R"," 17/40"," 23/60",38.3
--44,2018,12,"1D | 5R"," 6/20","4D | 9R"," 13/40"," 19/60",31.7
--43,2017,15,"1D | 5R"," 6/20","3D | 9R"," 12/40"," 18/60",30.0
--42,2016,11,"1D | 4R"," 5/20","2D | 10R"," 12/40"," 17/60",28.3
--41,2015,15,"1D | 4R"," 5/20","2D | 10R"," 12/40"," 17/60",28.3
--40,2014,17,"1D | 3R"," 4/20","2D | 10R"," 12/40"," 16/60",26.7
--39,2013,14,"1D | 3R"," 4/20","3D | 10R"," 13/40"," 17/60",28.3
--38,2012,21,"1D | 3R"," 4/20","3D | 7R"," 10/40"," 14/60",23.3
--37,2011,27,"1D | 3R"," 4/20","3D | 7R"," 10/40"," 14/60",23.3
--36,2010,34,"1D | 2R"," 3/20","3D | 6R"," 9/40"," 12/60",20.0
--35,2009,32,"1D | 2R"," 3/20","3D | 7R"," 10/40"," 13/60",21.7
--34,2008,28,"1D | 2R"," 3/20","5D | 5R"," 10/40"," 13/60",21.7
--33,2007,27,"1D | 2R"," 3/20","5D | 5R"," 10/40"," 13/60",21.7
--32,2006,34,"2D | 1R"," 3/20","4D | 4R"," 8/40"," 11/60",18.3
--31,2005,32,"2D | 1R"," 3/20","4D | 4R"," 8/40"," 11/60",18.3
--30,2004,32,"3D | 1R"," 4/20","3D | 5R"," 8/40"," 12/60",20.0
--29,2003,30,"3D | 1R"," 4/20","3D | 5R"," 8/40"," 12/60",20.0
--28,2002,32,"2D | 1R"," 3/20","4D | 4R"," 8/40"," 11/60",18.3
--27,2001,30,"2D | 2R"," 4/20","4D | 5R"," 9/40"," 13/60",21.7
--26,2000,33,"1D | 2R"," 3/20","4D | 5R"," 9/40"," 12/60",20.0
--25,1999,33,"1D | 2R"," 3/20","3D | 5R"," 8/40"," 11/60",18.3
--24,1998,43,"1D | 2R"," 3/20","2D | 4R"," 6/40"," 9/60",15.0
--23,1997,43,"1D | 2R"," 3/20","2D | 4R"," 6/40"," 9/60",15.0
--22,1996,21,"2D | 2R"," 4/20","5D | 5R | 1Ind"," 11/40"," 15/60",24.0
--21,1995,21,"2D | 2R"," 4/20","5D | 5R | 1Ind"," 11/40"," 15/60",24.0
--20,1994,23,"3D | 1R"," 4/20","6D | 4R"," 10/40"," 14/60",23.3
--19,1993,21,"3D | 1R"," 4/20","6D | 4R"," 10/40"," 14/60",23.3
--18,1992,15,"1D | 3R"," 4/20","6D | 5R"," 11/40"," 15/60",24.0
--17,1991,14,"1D | 3R"," 4/20","6D | 5R"," 11/40"," 15/60",24.0
--16,1990,16,"1D | 3R"," 4/20","5D | 4R"," 9/40"," 13/60",21.7
--15,1989,14,"1D | 3R"," 4/20","5D | 4R"," 9/40"," 13/60",21.7
--14,1988,14,"1D | 2R"," 3/20","5D | 4R"," 9/40"," 12/60",20.0
--13,1987,14,"1D | 2R"," 3/20","5D | 4R"," 9/40"," 12/60",20.0
--12,1986,15,"1D | 3R"," 4/20","4D | 3R"," 7/40"," 11/60",18.3
--11,1985,15,"1D | 3R"," 4/20","4D | 3R"," 7/40"," 11/60",18.3
--10,1984,22,"1D | 2R"," 3/20","2D | 3R"," 5/40"," 8/60",13.3
--9,1983,20,"1D | 2R"," 3/20","2D | 3R"," 5/40"," 8/60",13.3
--8,1982,,"1D| 1R",2/20,"3D | 1R"," 4/40"," 6/60",10.0
--7,1981,28,"1D | 1R"," 2/20","3D | 1R"," 4/40"," 6/60",10.0
--6,1980,,"1D | 1R"," 2/20","3D | 2R"," 5/40"," 7/60",11.7
--5,1979,14,"1D | 1R"," 2/20","3D | 2R"," 5/40"," 7/60",11.7
--4,1978,,"1D | 0R",1/20,"3D | 0R",3/40,4/60,6.7
--3,1977,28,"1D | 0R "," 1/20","3D | 0R"," 3/40"," 4/60",6.7
--2,1976,,"2D | 0R",2/20,"6D | 1R",7/40,9/60,15.0
--1,1975,5,"2D | 0R"," 2/20","6D | 1R"," 7/40"," 9/60",15.0
