--- python/tests/data/cawp_time/mock_territory_leg_tables/cawp_state_leg_60.csv
@@ -1,59 +1,0 @@
-time_period,state_fips,total_state_leg_count
-"1967","60",39
-"1968","60",39
-"1969","60",39
-"1970","60",39
-"1971","60",39
-"1972","60",39
-"1973","60",39
-"1974","60",39
-"1975","60",39
-"1976","60",39
-"1977","60",39
-"1978","60",39
-"1979","60",39
-"1980","60",39
-"1981","60",39
-"1982","60",39
-"1983","60",39
-"1984","60",39
-"1985","60",39
-"1986","60",39
-"1987","60",39
-"1988","60",39
-"1989","60",39
-"1990","60",39
-"1991","60",39
-"1992","60",39
-"1993","60",39
-"1994","60",39
-"1995","60",39
-"1996","60",39
-"1997","60",39
-"1998","60",39
-"1999","60",39
-"2000","60",39
-"2001","60",39
-"2002","60",39
-"2003","60",39
-"2004","60",39
-"2005","60",39
-"2006","60",39
-"2007","60",39
-"2008","60",39
-"2009","60",39
-"2010","60",39
-"2011","60",39
-"2012","60",39
-"2013","60",39
-"2014","60",39
-"2015","60",39
-"2016","60",39
-"2017","60",39
-"2018","60",39
-"2019","60",39
-"2020","60",39
-"2021","60",39
-"2022","60",39
-"2023","60",39
-"2024","60",39
