--- python/tests/data/cawp_time/mock_territory_leg_tables/cawp_state_leg_ZZ_territory.csv
@@ -1,2 +1,0 @@
-time_period,state_fips,total_state_leg_count
-"1999","ZZ",12345
